Move Kafka notification classes into new kafka package.
