+++ clients/notification-kafka/src/test/java/org/eclipse/hono/client/notification/kafka/NotificationKafkaConsumerConfigPropertiesTest.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.client.notification;
+package org.eclipse.hono.client.notification.kafka;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
